Update admin.php
additional delete promt

+++ attKris/admin.php
@@ -15,22 +15,20 @@
 <div id="main-content">
     <h2>All Records</h2>
     <?php
-        session_start();    
-        // $conn = mysqli_connect("localhost","root","","ce") or die("connection Failed");
-        require 'config.php';
-        $sql= "SELECT * FROM `students`";
+        $conn = mysqli_connect("localhost","root","","ce") or die("connection Failed");
+        $sql= "SELECT * FROM students";
         $result = mysqli_query($conn,$sql) or die("Query Unsucessful");
         if(mysqli_num_rows($result)>0){
-        ?>
+    ?>
     <table cellpadding="10px">
         <thead>
             <th>Roll No</th>
             <th>Name</th>
             <th>ID</th>
-            <!-- <th>Enrollment Year</th> -->
-            <!-- <th>Batch Year</th> -->
-            <!-- <th>Session No</th> -->
-            <!-- <th>Gender</th> -->
+            <th>Enrollment Year</th>
+            <th>Batch Year</th>
+            <th>Session No</th>
+            <th>Gender</th>
             <th>Division</th>
         </thead>
         <tbody>
@@ -38,17 +36,17 @@
             while($row = mysqli_fetch_assoc($result)){
             ?>
             <tr>
-                <td><?php echo $row['Roll No.']?></td>
-                <td><?php echo $row['ID']?></td>
-                <td><?php echo $row['Name']?></td>
-                <!-- <td><?php echo $row['Enrollment Year']?></td> -->
-                <!-- <td><?php echo $row['Batch Year']?></td> -->
-                <!-- <td><?php echo $row['Session No']?></td> -->
-                <!-- <td><?php echo $row['Gender']?></td> -->
-                <td><?php echo $row['Division']?></td>
+                <td><?php echo $row['Roll No.'] ?></td>
+                <td><?php echo $row['ID'] ?></td>
+                <td><?php echo $row['Name'] ?></td>
+                <td><?php echo $row['Enrollment Year'] ?></td>
+                <td><?php echo $row['Batch Year'] ?></td>
+                <td><?php echo $row['Session No'] ?></td>
+                <td><?php echo $row['Gender'] ?></td>
+                <td><?php echo $row['Division'] ?></td>
                 <td>
-                    <a href='edit.php?id=<?php echo $row['ID']?>'>Edit</a>
-                    <a href='delete.php?id=<?php echo $row['ID']?>'>Delete</a>
+                    <a href='edit.php?id=<?php echo $row['ID'] ?>'>Edit</a>
+                    <a href='delete.php?id=<?php echo $row['ID'] ?>' onclick="return confirmDelete('<?php echo $row['ID'] ?>')">Delete</a>
                 </td>
             </tr>
             <?php } ?>
@@ -58,4 +56,13 @@
 </div>
 </div>
 </body>
+<script>
+    function confirmDelete(id) {
+        var b = confirm("Do You Want To Delete The Data of " + id);
+        if (b == true) {
+            window.location.href = "delete.php?id=" + id;
+        }
+        return false;
+    }
+</script>
 </html>
